Fix intermediate deps for flavors other than debug.
Intermediate dependencies (e.g. R.class) were hard coded to the debug
build flavor. This commit fixes this by outputing the intermediate paths
for all flavors.

Note that there is no way to output only the dependencies for the flavor
being build at the moment. This script is executed before any particular
build (e.g. assembleDebug) is executed so we cannot know before hand
what deps to load. Fortunatelly most of the time, dependencies do not
conflict between flavors so loading them all should not be an issue
(most of the time).