Fix an issue when trying to resolve an non source file
It seems that in some cases the plugin is trying to resolve files who
are not part of the workspace, due to either their temporary nature, or
other IDE or tool file.

In my case, it was trying to resole files from the vim .undodir
directory, was failing to parse those files due to their binary nature,
and the whole sourceset resolution was failing in consequence.

To avoid that issue add a try catch around the file reading method to
ensure exception are not fired, and the resolution does not break.