Update the Gradle wrapper to 5.4.1
This commit updates the wrapper, but does not build correctly yet due to
the way Kotlin compiler libraries are resolved. To address this issue,
the previous mechanism of copying the dependent libraries into a local
folder from which they are included has to be updated. Since depending
on JAR files is non-ideal, a better solution would be to find the
(correctly packaged) libraries on a Maven repository and include them.