Add temporary source files
Temporary source files do not appear in SourcePath.all queries, but can
still be used for basic analysis (e.g. hovers). This is useful for
situations where the file does not belong onto the compilation classpath
(e.g. dependency source files which are already present on the
classpath) but should be analyzed anyway.