fix slow filewalk in some projects
As it is, with a `.gitignore`-file containing eg. `target`, it would ignore the folder `<the-repo>/target`, but not `<the-repo>/moduleA/target`. Added zero-or more glob dir-matcher `**` to fix this.

Also, .gitignore may have a `/` suffix, which will not match a Path to the same folder.

I tested: `defaultClassPathResolver(listOf(Paths.get("/path/to/problem/repo"))).classpath`

Before change: 1 min, 45 s

After change: 7 s

It's still possible to decrease this even further. Some ideas for:
- The code runs `mvn dependency:list -DincludeScope=test` for every `pom.xml` that is not in an ignored directory. For multimodule project, only the root-pom needs to be run, as this will find deps for all modules.
- mvn commands are usually a lot slower than parsing the xml-files manually, and as long as one does not have any 'used undecleared'-deps, that would be sufficient