Check if gradlew is executable
There are rare instances, where there is a non-executable "gradlew" next to a build script. In this case the language server will crash when trying to execute "gradlew".
The happens for example when using an openapi generator as this will place a "build.gradle" and a non-executable "gradlew" in the "build/generated" directory as helper files.