fix: Publish platform maven artifact
Publish a maven artifact for the java platform gradle module that is used to version dependencies across
other modules. Without this being published, users attempting to resolve the shared or server modules
in gradle from the github package registry will receive an error because the
"kotlin-language-server:platform" dependency referenced in the shared and server module POM's cannot be found.